Add files via upload
This project titled exploration and explanation is the second project of the udacity data analysis nanodegree program. The project instruction was to gather tweets of a particular twitter user with username WeRateDogs. This twitter user rate peoples dogs with humorous comments. In this project, I used Tweepy, a Python library, to query Twitter’s API for WeRateDogs Twitter data. The data will include retweet count and favorite count. I combined these two pieces of information with the default data provided by Udacity for this project.