Add files via upload
This project titled investigate a dataset is the first project on the udacity data analysis nanodegree program. In this project, i investigated a dataset that consist of information about over 100000  medical appointment in Brazil and is focused on the question of whether or not patients show up for their appointment and a number of characteristics about the patients who make the appointments.